ui corrections
