+++ lib/features/test_series/widgets/test_detail_widget.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hugeicons/hugeicons.dart';
 import 'package:iconify_flutter_plus/iconify_flutter_plus.dart';
-import 'package:iconify_flutter_plus/icons/mdi.dart';
 import 'package:zephyr/constants/app_constants.dart';
 
 class TestDetailWidget extends StatelessWidget {
