+++ lib/main.dart
@@ -3,7 +3,6 @@ import 'package:get/get.dart';
 import 'package:provider/provider.dart';
 import 'package:zephyr/common/provider/user_details_provider.dart';
 import 'package:zephyr/common/screens/splash_screen.dart';
-import 'package:zephyr/common/screens/test_screen.dart';
 import 'package:zephyr/common/service/notification_service.dart';
 import 'package:zephyr/constants/app_constants.dart';
 import 'package:zephyr/features/assignments/provider/assignment_provider.dart';
@@ -17,13 +16,7 @@ import 'package:zephyr/features/home/provider/home_page_provider.dart';
 import 'package:zephyr/features/live_class/provider/live_provider.dart';
 import 'package:zephyr/features/payment/provider/payment_provider.dart';
 import 'package:zephyr/features/test/provider/make_test_provider.dart';
-import 'package:zephyr/features/test/screens/make_your_test_solution.dart';
-import 'package:zephyr/features/test/screens/test_completion_screen.dart';
-import 'package:zephyr/features/test/screens/test_quiz_screen.dart';
 import 'package:zephyr/features/test_series/provider/test_series_provider.dart';
-import 'package:zephyr/features/test_series/screens/test_review_screen.dart';
-import 'package:zephyr/features/test_series/testetutor.dart';
-
 void main() async {
   // print("App starting");
   WidgetsFlutterBinding.ensureInitialized();
@@ -62,12 +55,6 @@ class MyApp extends StatelessWidget {
           useMaterial3: true,
           scaffoldBackgroundColor: AppColors.white),
       home: SplashScreen(),
-      // home: BottomNavScreen(),
-      // home: ChapterDetailsScreen(),
-      // home: ChapterAnalysisScreen(),
-      // home: TestQuizScreen(),
-      // home:  TestReviewScreen(),
-      // home: TestCompletionScreen(),
     );
   }
 }
